fix(go/adbc/driver/bigquery): Use number of rows (rather than schema) to check if we need an empty arrow iterator (#2674)
As the newly added tests shows, we might end up with a schema (`create
view`) and yet no rows in which case we would still have an issue #2173

This PR changes so we check if the number of rows is empty (and keep the
schema).

Note: The potential issue with using Schema vs. TotalRows is the same
(as the comment says for both of the fields they might not be
initialized in some cases before the first call to Next).